// plugins/nx-php/src/schemas/composer.schema.d.ts
export interface Package {
// ...
    'autoload-dev'?: {
        /**
         * This is an object of namespaces (keys) and the directories they can be found into (values, can be arrays of paths) by the autoloader.
         */
        'psr-0'?: {
            [k: string]: string | string[];
        };
        /**
         * This is an object of namespaces (keys) and the PSR-4 directories they can map to (values, can be arrays of paths) by the autoloader.
         */
        'psr-4'?: {
            [k: string]: string | string[];
        };
        /**
         * This is an array of paths that contain classes to be included in the class-map generation process.
         */
        classmap?: unknown[];
        /**
         * This is an array of files that are always required on every request.
         */
        files?: unknown[];
        [k: string]: unknown;
    };
// ...
    archive?: {
        /**
         * A base name for archive.
         */
        name?: string;
        /**
         * A list of patterns for paths to exclude or include if prefixed with an exclamation mark.
         */
        exclude?: unknown[];
        [k: string]: unknown;
    };
    /**
     * Composer options.
     */
    config?: {
// ...
        'archive-format'?: string;
        /**
         * The default archive path when not provided on cli, defaults to ".".
         */
        'archive-dir'?: string;
// ...
    };
// ...
    'scripts-aliases'?: {
        [k: string]: unknown[];
    };
    [k: string]: unknown;
}
// ...
export interface Autoload {
    /**
     * This is an object of namespaces (keys) and the directories they can be found in (values, can be arrays of paths) by the autoloader.
     */
    'psr-0'?: {
        [k: string]: string | string[];
    };
    /**
     * This is an object of namespaces (keys) and the PSR-4 directories they can map to (values, can be arrays of paths) by the autoloader.
     */
    'psr-4'?: {
        [k: string]: string | string[];
    };
    /**
     * This is an array of paths that contain classes to be included in the class-map generation process.
     */
    classmap?: unknown[];
    /**
     * This is an array of files that are always required on every request.
     */
    files?: unknown[];
    /**
     * This is an array of patterns to exclude from autoload classmap generation. (e.g. "exclude-from-classmap": ["/test/", "/tests/", "/Tests/"]
     */
    'exclude-from-classmap'?: unknown[];
    [k: string]: unknown;
}